Small modifications and bugfix?
The TP autosolver bug does not seem to be present in this version currently on GitHub. I did some testing, and this one has no problem in unity or in game, while the version on the steam workshop fails to solve in game. So I guess if you just use the script in this merge that will fix the bug.
Other changes:
- All LED materials are now unlit, to make it so the room going dark temporarily will not impair your vision and cause unwanted strikes
- Updated bundled manual to the latest on the repo